Fixed uint64->uint32 for millis related variables on thermostat branch

+++ examples/DSwithNotification/DSwithNotification.ino
@@ -114,7 +114,7 @@ void setup() {
 void loop() {
   SuplaDevice.iterate();
 
-  static uint64_t lastTime = 0;
+  static uint32_t lastTime = 0;
   if (millis() - lastTime > 60000) {
     lastTime = millis();
     static bool greetingsSend = false;
